+++ src/util/logger.ts
@@ -1,28 +1,42 @@
-import { ILogger } from 'src/platform/common/logger';
-import { LoggerImpl } from 'platform/loggerImpl';
-import { IConfiguration } from 'src/configuration/iconfiguration';
+import { LogOutputChannel, window } from 'vscode';
 
 export class Logger {
-  private static readonly cache = new Map<string, ILogger>();
-  private static configuration: IConfiguration | undefined = undefined;
+  private static output: LogOutputChannel | undefined;
+  private static readonly cache = new Map<string, Logger>();
 
-  static get(prefix: string): ILogger {
+  static get(prefix: string): Logger {
+    if (!this.output) {
+      this.output = window.createOutputChannel('Vim', { log: true });
+    }
+
     let logger = Logger.cache.get(prefix);
     if (logger === undefined) {
-      logger = LoggerImpl.get(prefix);
-      if (Logger.configuration) {
-        logger.configChanged(Logger.configuration);
-      }
+      logger = new Logger(prefix);
       Logger.cache.set(prefix, logger);
     }
 
     return logger;
   }
 
-  static configChanged(configuration: IConfiguration) {
-    Logger.configuration = configuration;
-    for (const logger of this.cache.values()) {
-      logger.configChanged(configuration);
-    }
+  private scope: string;
+
+  private constructor(scope: string) {
+    this.scope = scope;
+  }
+
+  error(msg: string): void {
+    Logger.output?.error(`[${this.scope}] ${msg}`);
+  }
+  warn(msg: string): void {
+    Logger.output?.warn(`[${this.scope}] ${msg}`);
+  }
+  info(msg: string): void {
+    Logger.output?.info(`[${this.scope}] ${msg}`);
+  }
+  debug(msg: string): void {
+    Logger.output?.debug(`[${this.scope}] ${msg}`);
+  }
+  trace(msg: string): void {
+    Logger.output?.trace(`[${this.scope}] ${msg}`);
   }
 }