+++ webpack.config.js
@@ -6,6 +6,7 @@ const path = require('path');
 const webpack = require('webpack');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
+const CircularDependencyPlugin = require('circular-dependency-plugin');
 
 /**@type {import('webpack').Configuration}*/
 const config = {
@@ -54,6 +55,19 @@ const config = {
       cleanOnceBeforeBuildPatterns: [], // disable initial clean
     }),
     new ForkTsCheckerWebpackPlugin(),
+    new CircularDependencyPlugin({
+      // exclude detection of files based on a RegExp
+      exclude: /node_modules/,
+      // include specific files based on a RegExp
+      include: /src/,
+      // add errors to webpack instead of warnings
+      failOnError: true,
+      // allow import cycles that include an asyncronous import,
+      // e.g. via import(/* webpackMode: "weak" */ './file.js')
+      allowAsyncCycles: true,
+      // set the current working directory for displaying module paths
+      cwd: process.cwd(),
+    }),
   ],
 };